feat(frontend): moved non-pages out of app, added 'use client's, expanded context to include villages and people
