+++ frontend/src/assets/globals.scss
