+++ frontend/src/components/RelationBrowser.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import type { TabsProps } from "antd";
 import { Tabs } from "antd";
